finished itinerary search

+++ src/main/java/com/example/quickescapes/service/HotelsService.java
@@ -4,10 +4,11 @@ import com.example.quickescapes.dao.Hotel;
 import com.example.quickescapes.dao.HotelQueries;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
 
 public interface HotelsService {
-     List<Hotel> searchHotels(Integer budget, LocalDate checkIn, LocalDate checkOut);
+     List<Hotel> searchHotels(BigDecimal budget, LocalDate checkIn, LocalDate checkOut);
 }
